Base: move Endpoint.exposed() tests into a separate spec file
